reorganise tracks, cars, weather content into content directory
