+++ cmd/server-manager/views/pages/content/tracks.html
