+++ handlers.go
@@ -130,58 +130,6 @@ func serverOptionsHandler(w http.ResponseWriter, r *http.Request) {
 	})
 }
 
-func carsHandler(w http.ResponseWriter, r *http.Request) {
-	cars, err := ListCars()
-
-	if err != nil {
-		logrus.Errorf("could not get car list, err: %s", err)
-		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-		return
-	}
-
-	ViewRenderer.MustLoadTemplate(w, r, "cars.html", map[string]interface{}{
-		"cars": cars,
-	})
-}
-
-func weatherHandler(w http.ResponseWriter, r *http.Request) {
-	weather, err := ListWeather()
-
-	if err != nil {
-		logrus.Errorf("could not get weather list, err: %s", err)
-		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-		return
-	}
-
-	ViewRenderer.MustLoadTemplate(w, r, "weather.html", map[string]interface{}{
-		"weathers": weather,
-	})
-}
-
-func apiCarUploadHandler(w http.ResponseWriter, r *http.Request) {
-	uploadHandler(w, r, "Car")
-}
-
-func apiWeatherUploadHandler(w http.ResponseWriter, r *http.Request) {
-	uploadHandler(w, r, "Weather")
-}
-
-func apiTrackUploadHandler(w http.ResponseWriter, r *http.Request) {
-	uploadHandler(w, r, "Track")
-}
-
-func tracksHandler(w http.ResponseWriter, r *http.Request) {
-	tracks, err := ListTracks()
-
-	if err != nil {
-		logrus.Errorf("could not get track list, err: %s", err)
-	}
-
-	ViewRenderer.MustLoadTemplate(w, r, "tracks.html", map[string]interface{}{
-		"tracks": tracks,
-	})
-}
-
 type contentFile struct {
 	Name     string `json:"name"`
 	FileType string `json:"type"`
@@ -268,155 +216,6 @@ func addFiles(files []contentFile, contentType string) error {
 	}
 
 	return nil
-}
-
-func trackDeleteHandler(w http.ResponseWriter, r *http.Request) {
-	trackName := mux.Vars(r)["name"]
-	tracksPath := filepath.Join(ServerInstallPath, "content", "tracks")
-
-	existingTracks, err := ListTracks()
-
-	if err != nil {
-		logrus.Errorf("could not get track list, err: %s", err)
-
-		AddFlashQuick(w, r, "couldn't get track list")
-
-		http.Redirect(w, r, r.Referer(), http.StatusFound)
-
-		return
-	}
-
-	var found bool
-
-	for _, track := range existingTracks {
-		if track.Name == trackName {
-			// Delete track
-			found = true
-
-			err := os.RemoveAll(filepath.Join(tracksPath, trackName))
-
-			if err != nil {
-				found = false
-				logrus.Errorf("could not remove track files, err: %s", err)
-			}
-
-			break
-		}
-	}
-
-	var message string
-
-	if found {
-		// confirm deletion
-		message = "Track successfully deleted!"
-	} else {
-		// inform track wasn't found
-		message = "Sorry, track could not be deleted."
-	}
-
-	AddFlashQuick(w, r, message)
-
-	http.Redirect(w, r, r.Referer(), http.StatusFound)
-}
-
-func weatherDeleteHandler(w http.ResponseWriter, r *http.Request) {
-	weatherKey := mux.Vars(r)["key"]
-	weatherPath := filepath.Join(ServerInstallPath, "content", "weather")
-
-	existingWeather, err := ListWeather()
-
-	if err != nil {
-		logrus.Errorf("could not get weather list, err: %s", err)
-
-		AddFlashQuick(w, r, "couldn't get weather list")
-
-		http.Redirect(w, r, r.Referer(), http.StatusFound)
-
-		return
-	}
-
-	var found bool
-
-	for key := range existingWeather {
-		if weatherKey == key {
-			// Delete car
-			found = true
-
-			err := os.RemoveAll(filepath.Join(weatherPath, weatherKey))
-
-			if err != nil {
-				found = false
-				logrus.Errorf("could not remove weather files, err: %s", err)
-			}
-
-			delete(existingWeather, key)
-
-			break
-		}
-	}
-
-	var message string
-
-	if found {
-		// confirm deletion
-		message = "Weather preset successfully deleted!"
-	} else {
-		// inform weather wasn't found
-		message = "Sorry, weather preset could not be deleted. Are you sure it was installed?"
-	}
-
-	AddFlashQuick(w, r, message)
-
-	http.Redirect(w, r, r.Referer(), http.StatusFound)
-}
-
-func carDeleteHandler(w http.ResponseWriter, r *http.Request) {
-	carName := mux.Vars(r)["name"]
-	carsPath := filepath.Join(ServerInstallPath, "content", "cars")
-
-	existingCars, err := ListCars()
-
-	if err != nil {
-		logrus.Errorf("could not get car list, err: %s", err)
-
-		AddFlashQuick(w, r, "couldn't get car list")
-
-		http.Redirect(w, r, r.Referer(), http.StatusFound)
-
-		return
-	}
-
-	var found bool
-
-	for _, car := range existingCars {
-		if car.Name == carName {
-			// Delete car
-			found = true
-
-			err := os.RemoveAll(filepath.Join(carsPath, carName))
-
-			if err != nil {
-				found = false
-				logrus.Errorf("could not remove car files, err: %s", err)
-			}
-
-			break
-		}
-	}
-
-	var message string
-
-	if found {
-		// confirm deletion
-		message = "Car successfully deleted!"
-	} else {
-		// inform car wasn't found
-		message = "Sorry, car could not be deleted. Are you sure it was installed?"
-	}
-
-	AddFlashQuick(w, r, message)
-
-	http.Redirect(w, r, r.Referer(), http.StatusFound)
 }
 
 func getSession(r *http.Request) *sessions.Session {
